@@ -1,8 +1,9 @@
 import ast
+import contextlib
 import json
 from copy import deepcopy
 from datetime import timezone
-from typing import Any, Dict, List, Union
+from typing import Any, Callable, Dict, List, Union
 
 from dateutil import parser
 from dateutil.tz import gettz
@@ -66,7 +67,7 @@ class LogEntryManager(models.Manager):
                 kwargs.setdefault("additional_data", get_additional_data())
 
             # Delete log entries with the same pk as a newly created model.
-            # This should only be necessary when an pk is used twice.
+            # This should only be necessary when a pk is used twice.
             if kwargs.get("action", None) is LogEntry.Action.CREATE:
                 if (
                     kwargs.get("object_id", None) is not None
@@ -110,7 +111,7 @@ class LogEntryManager(models.Manager):
         from auditlog.cid import get_cid
 
         pk = self._get_pk_value(instance)
-        if changed_queryset is not None:
+        if changed_queryset:
             kwargs.setdefault(
                 "content_type", ContentType.objects.get_for_model(instance)
             )
@@ -225,7 +226,7 @@ class LogEntryManager(models.Manager):
         pk_field = instance._meta.pk.name
         pk = getattr(instance, pk_field, None)
 
-        # Check to make sure that we got an pk not a model object.
+        # Check to make sure that we got a pk not a model object.
         if isinstance(pk, models.Model):
             pk = self._get_pk_value(pk)
         return pk
@@ -273,7 +274,10 @@ class LogEntryManager(models.Manager):
         for field in instance_copy._meta.fields:
             if not field.is_relation:
                 value = getattr(instance_copy, field.name)
-                setattr(instance_copy, field.name, field.to_python(value))
+                try:
+                    setattr(instance_copy, field.name, field.to_python(value))
+                except ValidationError:
+                    continue
         return instance_copy
 
     def _get_applicable_model_fields(
@@ -355,6 +359,7 @@ class LogEntry(models.Model):
     action = models.PositiveSmallIntegerField(
         choices=Action.choices, verbose_name=_("action"), db_index=True
     )
+    changes_text = models.TextField(blank=True, verbose_name=_("change message"))
     changes = models.JSONField(null=True, verbose_name=_("change message"))
     actor = models.ForeignKey(
         to=settings.AUTH_USER_MODEL,
@@ -408,7 +413,7 @@ class LogEntry(models.Model):
         """
         :return: The changes recorded in this log entry as a dictionary object.
         """
-        return self.changes or {}
+        return changes_func(self)
 
     @property
     def changes_str(self, colon=": ", arrow=" \u2192 ", separator="; "):
@@ -456,7 +461,7 @@ class LogEntry(models.Model):
                 changes_display_dict[field_name] = values
                 continue
             values_display = []
-            # handle choices fields and Postgres ArrayField to get human readable version
+            # handle choices fields and Postgres ArrayField to get human-readable version
             choices_dict = None
             if getattr(field, "choices", []):
                 choices_dict = dict(field.choices)
@@ -543,7 +548,7 @@ class AuditlogHistoryField(GenericRelation):
     A subclass of py:class:`django.contrib.contenttypes.fields.GenericRelation` that sets some default
     variables. This makes it easier to access Auditlog's log entries, for example in templates.
 
-    By default this field will assume that your primary keys are numeric, simply because this is the most
+    By default, this field will assume that your primary keys are numeric, simply because this is the most
     common case. However, if you have a non-integer primary key, you can simply pass ``pk_indexable=False``
     to the constructor, and Auditlog will fall back to using a non-indexed text based field for this model.
 
@@ -552,14 +557,12 @@ class AuditlogHistoryField(GenericRelation):
 
     :param pk_indexable: Whether the primary key for this model is not an :py:class:`int` or :py:class:`long`.
     :type pk_indexable: bool
-    :param delete_related: By default, including a generic relation into a model will cause all related
-        objects to be cascade-deleted when the parent object is deleted. Passing False to this overrides this
-        behavior, retaining the full auditlog history for the object. Defaults to True, because that's
-        Django's default behavior.
+    :param delete_related: Delete referenced auditlog entries together with the tracked object.
+        Defaults to False to keep the integrity of the auditlog.
     :type delete_related: bool
     """
 
-    def __init__(self, pk_indexable=True, delete_related=True, **kwargs):
+    def __init__(self, pk_indexable=True, delete_related=False, **kwargs):
         kwargs["to"] = LogEntry
 
         if pk_indexable:
@@ -582,3 +585,24 @@ class AuditlogHistoryField(GenericRelation):
         # method.  However, because we don't want to delete these related
         # objects, we simply return an empty list.
         return []
+
+
+# should I add a signal receiver for setting_changed?
+changes_func = None
+
+
+def _changes_func() -> Callable[[LogEntry], Dict]:
+    def json_then_text(instance: LogEntry) -> Dict:
+        if instance.changes:
+            return instance.changes
+        elif instance.changes_text:
+            with contextlib.suppress(ValueError):
+                return json.loads(instance.changes_text)
+        return {}
+
+    def default(instance: LogEntry) -> Dict:
+        return instance.changes or {}
+
+    if settings.AUDITLOG_USE_TEXT_CHANGES_IF_JSON_IS_NOT_PRESENT:
+        return json_then_text
+    return default
