@@ -21,7 +21,7 @@ class SimpleModel(models.Model):
     integer = models.IntegerField(blank=True, null=True)
     datetime = models.DateTimeField(auto_now=True)
 
-    history = AuditlogHistoryField()
+    history = AuditlogHistoryField(delete_related=True)
 
     def __str__(self):
         return self.text
@@ -39,7 +39,7 @@ class AltPrimaryKeyModel(models.Model):
     integer = models.IntegerField(blank=True, null=True)
     datetime = models.DateTimeField(auto_now=True)
 
-    history = AuditlogHistoryField(pk_indexable=False)
+    history = AuditlogHistoryField(delete_related=True, pk_indexable=False)
 
 
 class UUIDPrimaryKeyModel(models.Model):
@@ -54,7 +54,7 @@ class UUIDPrimaryKeyModel(models.Model):
     integer = models.IntegerField(blank=True, null=True)
     datetime = models.DateTimeField(auto_now=True)
 
-    history = AuditlogHistoryField(pk_indexable=False)
+    history = AuditlogHistoryField(delete_related=True, pk_indexable=False)
 
 
 class ProxyModel(SimpleModel):
@@ -84,7 +84,7 @@ class RelatedModel(RelatedModelParent):
         to="SimpleModel", on_delete=models.CASCADE, related_name="reverse_one_to_one"
     )
 
-    history = AuditlogHistoryField()
+    history = AuditlogHistoryField(delete_related=True)
 
 
 class ManyRelatedModel(models.Model):
@@ -95,7 +95,7 @@ class ManyRelatedModel(models.Model):
     recursive = models.ManyToManyField("self")
     related = models.ManyToManyField("ManyRelatedOtherModel", related_name="related")
 
-    history = AuditlogHistoryField()
+    history = AuditlogHistoryField(delete_related=True)
 
     def get_additional_data(self):
         related = self.related.first()
@@ -107,7 +107,7 @@ class ManyRelatedOtherModel(models.Model):
     A model related to ManyRelatedModel as many-to-many.
     """
 
-    history = AuditlogHistoryField()
+    history = AuditlogHistoryField(delete_related=True)
 
 
 @auditlog.register(include_fields=["label"])
@@ -119,7 +119,7 @@ class SimpleIncludeModel(models.Model):
     label = models.CharField(max_length=100)
     text = models.TextField(blank=True)
 
-    history = AuditlogHistoryField()
+    history = AuditlogHistoryField(delete_related=True)
 
 
 class SimpleExcludeModel(models.Model):
@@ -130,7 +130,7 @@ class SimpleExcludeModel(models.Model):
     label = models.CharField(max_length=100)
     text = models.TextField(blank=True)
 
-    history = AuditlogHistoryField()
+    history = AuditlogHistoryField(delete_related=True)
 
 
 class SimpleMappingModel(models.Model):
@@ -142,7 +142,7 @@ class SimpleMappingModel(models.Model):
     vtxt = models.CharField(verbose_name="Version", max_length=100)
     not_mapped = models.CharField(max_length=100)
 
-    history = AuditlogHistoryField()
+    history = AuditlogHistoryField(delete_related=True)
 
 
 @auditlog.register(mask_fields=["address"])
@@ -154,7 +154,7 @@ class SimpleMaskedModel(models.Model):
     address = models.CharField(max_length=100)
     text = models.TextField()
 
-    history = AuditlogHistoryField()
+    history = AuditlogHistoryField(delete_related=True)
 
 
 class AdditionalDataIncludedModel(models.Model):
@@ -167,7 +167,7 @@ class AdditionalDataIncludedModel(models.Model):
     text = models.TextField(blank=True)
     related = models.ForeignKey(to=SimpleModel, on_delete=models.CASCADE)
 
-    history = AuditlogHistoryField()
+    history = AuditlogHistoryField(delete_related=True)
 
     def get_additional_data(self):
         """
@@ -194,7 +194,7 @@ class DateTimeFieldModel(models.Model):
     time = models.TimeField()
     naive_dt = models.DateTimeField(null=True, blank=True)
 
-    history = AuditlogHistoryField()
+    history = AuditlogHistoryField(delete_related=True)
 
 
 class ChoicesFieldModel(models.Model):
@@ -216,7 +216,7 @@ class ChoicesFieldModel(models.Model):
     status = models.CharField(max_length=1, choices=STATUS_CHOICES)
     multiplechoice = models.CharField(max_length=255, choices=STATUS_CHOICES)
 
-    history = AuditlogHistoryField()
+    history = AuditlogHistoryField(delete_related=True)
 
 
 class CharfieldTextfieldModel(models.Model):
@@ -229,7 +229,7 @@ class CharfieldTextfieldModel(models.Model):
     longchar = models.CharField(max_length=255)
     longtextfield = models.TextField()
 
-    history = AuditlogHistoryField()
+    history = AuditlogHistoryField(delete_related=True)
 
 
 class PostgresArrayFieldModel(models.Model):
@@ -251,7 +251,7 @@ class PostgresArrayFieldModel(models.Model):
         models.CharField(max_length=1, choices=STATUS_CHOICES), size=3
     )
 
-    history = AuditlogHistoryField()
+    history = AuditlogHistoryField(delete_related=True)
 
 
 class NoDeleteHistoryModel(models.Model):
@@ -302,6 +302,29 @@ class SerializeNaturalKeyRelatedModel(models.Model):
     value = models.IntegerField()
 
     history = AuditlogHistoryField(delete_related=False)
+
+
+class SimpleNonManagedModel(models.Model):
+    """
+    A simple model with no special things going on.
+    """
+
+    text = models.TextField(blank=True)
+    boolean = models.BooleanField(default=False)
+    integer = models.IntegerField(blank=True, null=True)
+    datetime = models.DateTimeField(auto_now=True)
+
+    history = AuditlogHistoryField(delete_related=True)
+
+    def __str__(self):
+        return self.text
+
+    class Meta:
+        managed = False
+
+
+class AutoManyRelatedModel(models.Model):
+    related = models.ManyToManyField(SimpleModel)
 
 
 auditlog.register(AltPrimaryKeyModel)
